requirements.txt
pandas
github-copilot-sdk
openpyxl
